Update .SRCINFO
Sorry for the dumb commits, I'm trying to reason about what to do with a
VCS package (*-git) when the upstream hasn't changed but the packaging
in here has.  For some reason, the man page isn't installed when I
install it either.  If you read this, don't hesitate to get in touch
with me to share your guidance :)